Extract common controller utils to a separate file
This is better than cross-importing between controllers and provides
a single place for additional utils to be placed.
