+++ .vscode/settings.json
@@ -1,5 +1,5 @@
 {
-    "python.autoComplete.extraPaths": ["stubs"],
+    "python.autoComplete.extraPaths": ["modules", "stubs"],
     "python.linting.flake8Enabled": true,
     "python.linting.mypyEnabled": true,
     "python.pythonPath": "${env:VIRTUAL_ENV}/bin/python",
