+++ setup.cfg
@@ -14,6 +14,8 @@ ignore =
     # don't require commas in places that only Python 3.6 requires them (we
     # support Python 3.5)
     C816
+    # Ignore flake8's complaints about import positions since we're using isort
+    E402
     # W503 and W504 conflict; ignore the one that is less visually appealing.
     W504
 
@@ -49,7 +51,7 @@ not_skip =
 
 
 [mypy]
-mypy_path = stubs
+mypy_path = modules, stubs
 
 disallow_any_explicit = True
 disallow_any_generics = True
